+++ tests/unit/configuration_test.py
@@ -2151,7 +2151,8 @@ class TestConfiguration(TestConfigurationBase):
             )
             self.assertRaisesRegexp(BadConfiguration, expected_msg, config.parse)
 
-    @skipIf(sys.version_info < (2, 7, 0), "Skipping tests under Python 2.6")
+    @skipIf(sys.version_info < (3, 5, 0), "Skipping tests under Python < 3.5")
+    @skipIf(platform.system() == "Darwin", "Skipping under OSX")
     def test_set_json_library_on_apply_config(self):
         current_json_lib = scalyr_util.get_json_lib()
         self.assertEqual(current_json_lib, "json")
@@ -2171,11 +2172,11 @@ class TestConfiguration(TestConfigurationBase):
         new_json_lib = scalyr_util.get_json_lib()
         self.assertEqual(new_json_lib, "json")
 
-        # auth should fall back to ujson again
+        # auth should fall back to orjson again
         self._write_file_with_separator_conversion(
             """{
              api_key: "hi there",
-            json_library: "ujson"
+            json_library: "orjson"
           }
         """
         )
@@ -2185,7 +2186,7 @@ class TestConfiguration(TestConfigurationBase):
         config.apply_config()
 
         new_json_lib = scalyr_util.get_json_lib()
-        self.assertEqual(new_json_lib, "ujson")
+        self.assertEqual(new_json_lib, "orjson")
 
     @skipIf(sys.version_info < (2, 7, 0), "Skipping tests under Python 2.6")
     def test_apply_config_without_parse(self):